delete(): don't allow None documents in list of IDs
Doesn't make sense to delete None documents, so make it an exception.
This means we need to filter the original list for not Nones.
Not sure an exception is the wanted behaviour here, but at the same time
I wanted to avoid the '<delete></delete>' command being sent over to Solr
(although that does NOT delete all documents).