Refactor start-test-solr.sh to support Solr 4 and Solr 5
The test suite now uses a single-core example that's present in both
Solr 4 and 5. The script detects the appropriate Solr version at
runtime and prepares a Solr home directory accordingly. Some of the
tests, which assumed the core is called "core0," have been modified
to use a core called "collection1."

To be clear: the test suite runs with Solr 5, but it doesn't yet pass.