Tests: launch Solr instances on different ports
unittest.TestCase.setUpClass will call the super-class for
every method which isn’t explicitly overridden in the base
class. This means that the call sequence looks something like this:

SolrCloudTestCase.setUpClass -> start SolrCloud
SolrTestCase.setUpClass -> start Solr
SolrTestCase.tearDownClass -> kill everything
SolrCoreAdminTestCase.setUpClass -> start Solr
… etc …

Besides adding minutes to the test time even on a machine
with a fast SSD, this makes the test process inherently
unstable because the order in which everything happens
may change randomly depending on the test selection. This is
particularly catastrophic with Kazoo, which attempts to
maintain persistent connections using an `atexit` hack which
cannot be cleanly shut down when the server is randomly
killed by another test.

This commit changes that to simply start both flavors of
Solr on different ports when the test runner starts and
tear them down once at the end.