chg: remove __del__ methods
The __del__ methods were added in an attempt to avoid Kazoo-related
failures as part of the SolrCloud support but can cause other problems
on different versions of Python (see #193).

Since the errors in question were observed during testing this commit
removes the __del__ methods and we will have to find an alternative for
making tests fail safely.