GH#151: delete(id=<list>) assumed no parent/child docs
Rearrange the multiple-id delete() test code according to the new default
document test set.

Chose to filter out the parent/child documents from this specific part of the
test, under the assumption that Solr will behave correctly whether the id
to be deleted belongs to a leaf, parent, child or grandchild document.

The test code should be generic enough to not need modifications when
the document test set (self.docs) is expanded.