Fixed: Different JSON and XML behaviour
This commit adds a _build_json_doc method that strips keys with empty values from the doc prior to being serialized to JSON. This behaviour matches the behaviour of the XML method. It also uses the same method (_is_null_value) so if this behaviour needs to change in the future it should remain consistent.

It includes a new test to ensure the behaviours of the two methods remain consistent.

Fixes #321